backend/requirements.txt
alembic
fastapi[standard]
psycopg2
psycopg2-binary
pyjwt
requests
sqlalchemy[asyncio]
uvicorn[standard]